event/requirements.txt
Django
Pillow
qrcode
reportlab
python-decouple
requests
django-crispy-forms
crispy-bootstrap5
django-sslserver
python-dotenv
celery
redis
django-celery-beat
django-celery-results
